requirements.txt
pytesseract
opencv-python
Pillow
deep-translator
